+++ CHANGELOG.md
@@ -6,9 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Remove item index in `addToCart` mutation.
 
-## [0.57.0] - 2021-03-23
-
+## [0.57.0] - 2021-03-23 [YANKED]
 ### Added
 - Parameter `allowOutdatedData` to `addToCart` mutation.